chore: relocate utils tests
