+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking Changes
 
+* [BREAKING][removal][rust] `tokens_to_base_units`, `base_units_to_tokens` and `TokenParseError` are removed from `miden_client::utils`. They format and parse fungible amounts for display against a faucet's decimals, which only the CLI needs, so they now live in the CLI crate ([#2515](https://github.com/0xMiden/rust-sdk/pull/2515)).
 * [BREAKING][rust] The workspace MSRV and the pinned toolchain are raised to Rust 1.98. Building the client now requires a 1.98 or newer compiler. The declared MSRV of 1.96 could not build the locked dependency graph, whose Miden packages require 1.96.1 ([#2471](https://github.com/0xMiden/rust-sdk/issues/2471)).
 
 ### Enhancements
